renames websocket to user_session
