+++ services/web/server/src/simcore_service_webserver/projects/_observer.py
@@ -14,7 +14,7 @@ from servicelib.aiohttp.observer import (
 from servicelib.utils import logged_gather
 
 from ..notifications import project_logs
-from ..resource_manager.websocket_manager import PROJECT_ID_KEY, managed_resource
+from ..resource_manager.user_sessions import PROJECT_ID_KEY, managed_resource
 from .projects_api import retrieve_and_notify_project_locked_state
 
 _logger = logging.getLogger(__name__)
