+++ services/web/server/src/simcore_service_webserver/socketio/messages.py
@@ -12,7 +12,7 @@ from servicelib.json_serialization import json_dumps
 from servicelib.utils import fire_and_forget_task, logged_gather
 from socketio import AsyncServer
 
-from ..resource_manager.websocket_manager import managed_resource
+from ..resource_manager.user_sessions import managed_resource
 from ._utils import get_socket_server
 
 _logger = logging.getLogger(__name__)
